Use run! instead of run when calling theme-check
Whereas theme-check is designed to be used as a CLI, we're directly
calling the code under the hood. The #run method we currently use is not
really designed for our use case, as it includes `exit` which we don't
want a dependency to do for us.

This commit switches to using #run! because that will at worst raise an
error which we can bubble up in CLI-appropriate form.